requirements.txt
argparse
CairoSVG
Pillow
